HSEARCH-3759 Move RootFailureCollector to SPI
We'll need it in the ORM mapper when we expose schema management
operations, in particular validation.
